Fix some issues prevents recursion in submodules.
Now get_modules works if there are no submodules, and stop throwing
exceptions for unusal submodule urls, since odeint has an external
documentation submodule.

+++ common/code/boost_super_project.php
@@ -14,23 +14,27 @@ class BoostSuperProject {
 
     public function parse_config_file($path) {
         if ($this->git_branch) {
+            $blob = $this->run_git("ls-tree {$this->git_branch} \"{$path}\"");
+            if (!$blob || !$blob[0]) {
+                return array();
+            }
+            $blob = preg_split("@[\t ]@", $blob[0]);
+            $blob = $blob[2];
+
             if (self::git_version() >= array(1,8,4,0)) {
-                $blob = $this->run_git("ls-tree {$this->git_branch} \"{$path}\"");
-                $blob = preg_split("@[\t ]@", $blob[0]);
-                $blob = $blob[2];
                 return $this->run_git("config -l --blob {$blob}");
             }
             else {
                 $temp_file = tempnam(sys_get_temp_dir(), 'boost-git-');
                 file_put_contents($temp_file, implode("\n",
-                    $this->run_git("show \"{$this->git_branch}:{$path}\"")));
+                    $this->run_git("show {$blob}")));
                 $result = $this->run_git("config -l -f \"{$temp_file}\"");
                 unlink($temp_file);
                 return $result;
             }
         }
         else {
-            return $this->run_git("config -l -f \"{$path}\"");
+            return is_file($path) ? $this->run_git("config -l -f \"{$path}\"") : array();
         }
     }
 
@@ -49,16 +53,21 @@ class BoostSuperProject {
             }
         }
 
+        $result = array();
         foreach ($modules as $name => $settings) {
             if (empty($settings['url'])) {
                 throw new RuntimeException("Missing URL for {$name}.");
             }
             else if (!preg_match('@^\.\./(\w+)\.git$@', $settings['url'])) {
-                throw new RuntimeException("Invalid URL for {$name}.");
+                //throw new RuntimeException("Invalid URL for {$name}.");
+                echo "Ignoring submodule '{$name}' in '{$this->location}'.\n";
+            }
+            else {
+                $result[$name] = $settings;
             }
         }
 
-        return $modules;
+        return $result;
     }
 
     public function run_git($command) {
